General Interface for Cox Models

cox_reg() is a way to generate a specification of a model before fitting and allows the model to be created using different packages in R. The main arguments for the model are:

penalty: The total amount of regularization in the model. Note that this must be zero for some engines.

mixture: The mixture amounts of different types of regularization (see below). Note that this will be ignored for some engines.

These arguments are converted to their specific names at the time that the model is fit. Other options and arguments can be set using set_engine(). If left to their defaults here (NULL), the values are taken from the underlying model functions. If parameters need to be modified, update() can be used in lieu of recreating the object from scratch.

cox_reg(mode = "censored regression", penalty = NULL, mixture = NULL)

# S3 method for cox_reg
update(
  object,
  parameters = NULL,
  penalty = NULL,
  mixture = NULL,
  fresh = FALSE,
  ...
)

Arguments

mode
A single character string for the type of model. Possible values for this model are "unknown", or "censored regression".

penalty
A non-negative number representing the total amount of regularization (glmnet, keras, and spark only). For keras models, this corresponds to purely L2 regularization (aka weight decay) while the other models can be a combination of L1 and L2 (depending on the value of mixture; see below).

mixture
A number between zero and one (inclusive) that is the proportion of L1 regularization (i.e. lasso) in the model. When mixture = 1, it is a pure lasso model while mixture = 0 indicates that ridge regression is being used. (glmnet and spark only).

object
A cox model specification.

parameters
A 1-row tibble or named list with main parameters to update. If the individual arguments are used, these will supersede the values in parameters. Also, using engine arguments in this object will result in an error.

fresh
A logical for whether the arguments should be modified in-place of or replaced wholesale.

...
Not used for update().

Details

For cox_reg(), the mode will always be "survival".

The model can be created using the fit() function using the following engines:

R: "survival" (the default)

Engine Details

Engines may have pre-set default arguments when executing the model fit call. For this type of model, the template of the fit calls are:

survival engine

survival::coxph(formula = missing_arg())

Examples

library(survival)

cox_mod <-
  cox_reg() %>%
  set_engine("survival") %>%
  fit(Surv(time, status) ~ x, data = aml)
model <- cox_reg(penalty = 10, mixture = 0.1)
model

#> Cox Model Specification (censored regression)
#> 
#> Main Arguments:
#>   penalty = 10
#>   mixture = 0.1
#> 

update(model, penalty = 1)

#> Cox Model Specification (censored regression)
#> 
#> Main Arguments:
#>   penalty = 1
#>   mixture = 0.1
#> 

update(model, penalty = 1, fresh = TRUE)

#> Cox Model Specification (censored regression)
#> 
#> Main Arguments:
#>   penalty = 1
#>