Ensembles of decision trees

ipred::bagging() fits an ensemble of decision trees.

Details

For this engine, there is a single mode: censored regression

Tuning Parameters

This model has no tuning parameters.

Translation from parsnip to the original package (censored regression)

bag_tree(engine = "ipred") %>% 
  set_mode("censored regression") %>% 
  translate()

## Bagged Decision Tree Model Specification (censored regression)
## 
## Main Arguments:
##   cost_complexity = 0
##   min_n = 2
## 
## Computational engine: ipred 
## 
## Model fit template:
## ipred::bagging(formula = missing_arg(), data = missing_arg())

Preprocessing requirements

This engine does not require any special encoding of the predictors. Categorical predictors can be partitioned into groups of factor levels (e.g. {a, c} vs {b, d}) when splitting at a node. Dummy variables are not required for this model.

References

Breiman L. 1996. “Bagging predictors”. Machine Learning. 24 (2): 123-140

Hothorn T, Lausen B, Benner A, Radespiel-Troeger M. 2004. Bagging Survival Trees. Statistics in Medicine, 23(1), 77–91.

Kuhn, M, and K Johnson. 2013. Applied Predictive Modeling. Springer.