Boosted trees via mboost

blackboost_train() is a wrapper for the blackboost() function in the mboost package that fits tree-based models where all of the model arguments are in the main function.

Usage

blackboost_train(
  formula,
  data,
  family,
  weights = NULL,
  teststat = "quad",
  testtype = "Teststatistic",
  mincriterion = 0,
  minsplit = 10,
  minbucket = 4,
  maxdepth = 2,
  saveinfo = FALSE,
  ...
)

Arguments

teststat

a character specifying the type of the test statistic to be applied for variable selection.

testtype

a character specifying how to compute the distribution of the test statistic. The first three options refer to p-values as criterion, Teststatistic uses the raw statistic as criterion. Bonferroni and Univariate relate to p-values from the asymptotic distribution (adjusted or unadjusted). Bonferroni-adjusted Monte-Carlo p-values are computed when both Bonferroni and MonteCarlo are given.

mincriterion

the value of the test statistic or 1 - p-value that must be exceeded in order to implement a split.

minsplit

the minimum sum of weights in a node in order to be considered for splitting.

minbucket

the minimum sum of weights in a terminal node.

maxdepth

maximum depth of the tree. The default maxdepth = Inf means that no restrictions are applied to tree sizes.

saveinfo

logical. Store information about variable selection procedure in info slot of each partynode.

...

Other arguments to pass.

x

A data frame or matrix of predictors.

y

A factor vector with 2 or more levels

Value

A fitted blackboost model.

Examples

blackboost_train(Surv(time, status) ~ age + ph.ecog,
  data = lung[-14, ], family = mboost::CoxPH()
)
#> 
#> 	 Model-based Boosting
#> 
#> Call:
#> mboost::blackboost(formula = formula, data = data, family = family,     control = mboost::boost_control(), tree_controls = partykit::ctree_control(teststat = "quad",         testtype = "Teststatistic", mincriterion = 0, minsplit = 10,         minbucket = 4, maxdepth = 2, saveinfo = FALSE))
#> 
#> 
#> 	 Cox Partial Likelihood 
#> 
#> Loss function:  
#> 
#> Number of boosting iterations: mstop = 100 
#> Step size:  0.1 
#> Offset:  0 
#> Number of baselearners:  1 
#>