Parsnip Engines for Parametric Survival Models

These are engines for the parsnip::survival_reg() model specification.

survival_reg() is a way to generate a specification of a model before fitting and allows the model to be created using R. The main argument for the model is:

dist: The probability distribution of the outcome.

This argument is converted to its specific names at the time that the model is fit. Other options and argument can be set using set_engine(). If left to its default here (NULL), the value is taken from the underlying model functions.

Since survival models typically involve censoring (and require the use of survival::Surv() objects), the fit() function will require that the survival model be specified via the formula interface.

Also, for the flexsurv::flexsurvfit engine, the typical strata function cannot be used. To achieve the same effect, the extra parameter roles can be used (as described above).

The model can be created using the fit() function using the following engines:

R: "flexsurv", "survival" (the default)

Engine Details

Engines may have pre-set default arguments when executing the model fit call. For this type of model, the template of the fit calls are below.

flexsurv

survival_reg() %>% 
  set_engine("flexsurv") %>% 
  set_mode("censored regression") %>% 
  translate()

## Parametric Survival Regression Model Specification (censored regression)
## 
## Computational engine: flexsurv 
## 
## Model fit template:
## flexsurv::flexsurvreg(formula = missing_arg(), data = missing_arg(), 
##     weights = missing_arg())

survival

survival_reg() %>% 
  set_engine("survival") %>% 
  set_mode("censored regression") %>% 
  translate()

## Parametric Survival Regression Model Specification (censored regression)
## 
## Computational engine: survival 
## 
## Model fit template:
## survival::survreg(formula = missing_arg(), data = missing_arg(), 
##     weights = missing_arg(), model = TRUE)

Note that model = TRUE is needed to produce quantile predictions when there is a stratification variable and can be overridden in other cases.

fit() passes the data directly to survival::curvreg() so that its formula method can create dummy variables as-needed.

Parameter translations

The standardized parameter names in parsnip can be mapped to their original names in each engine that has main parameters. Each engine typically has a different default value (shown in parentheses) for each parameter.

parsnip	survival	flexsurv

dist	dist	dist

References

Jackson, C. (2016). flexsurv: A Platform for Parametric Survival Modeling in R. Journal of Statistical Software, 70(8), 1 - 33.

See also